[ALG] move algorithm prim types/accessor to a new place
